single cell browser
not including ld tracks
