+++ src/genome-browser/tracks/atac-seq-peaks.ts
@@ -1,0 +1,38 @@
+import { TrackList } from "../types";
+
+export const AtackSeqPeaksTracks: TrackList = {
+  "ATAC Seq Peaks": [
+    {
+      title: "Astrocytes",
+      url: "https://downloads.wenglab.org/Astro.PeakCalls.bb",
+    },
+    {
+      title: "Endothelial Cells",
+      url: "https://downloads.wenglab.org/Endo.PeakCalls.bb",
+    },
+    {
+      title: "Oligodendrocyte Precursor Cells",
+      url: "https://downloads.wenglab.org/OPC.PeakCalls.bb",
+    },
+    {
+      title: "Excitatory Neurons",
+      url: "https://downloads.wenglab.org/Exc.PeakCalls.bb",
+    },
+    {
+      title: "Oligodendrocytes",
+      url: "https://downloads.wenglab.org/Oligo.PeakCalls.bb",
+    },
+    {
+      title: "Inhibitory Neurons",
+      url: "https://downloads.wenglab.org/Inh.PeakCalls.bb",
+    },
+    {
+      title: "Microglia",
+      url: "https://downloads.wenglab.org/Micro.PeakCalls.bb",
+    },
+    {
+      title: "All CellTypes",
+      url: "https://downloads.wenglab.org/All.celltypes.Union.PeakCalls.bb",
+    },
+  ],
+};
