+++ src/genome-browser/tracks/conservation.ts
@@ -3,7 +3,8 @@ import { TrackList } from "../types";
 export const EvoConservationTracks: TrackList = {
   "Conservation Tracks": [
     {
-      title: "240-mammal phyloP conservation score (Vertical Viewing Range [-2 to 9])",
+      title:
+        "240-mammal phyloP conservation score (Vertical Viewing Range [-2 to 9])",
       url: "https://downloads.wenglab.org/241-mammalian-2020v2.bigWig",
     },
     {
